… libhb/comb_detect.c
… libhb/nlmeans.c
… libhb/taskset.c