Patch to improve filters scaling on many threads. 20% perf boost on a 28 core x86.

**Test on:**

- [x] Windows 10+  (via MinGW)
- [ ] macOS 10.13+
- [ ] Ubuntu Linux