@jstebbins 2021-08-10
Big patch.  Going to take some time to figure out what they did.

Did they supply any high level overview of *how* this improvement works?  If not, can you request one?