@galad87 2021-08-10
I can try to request.
From what I see before there was a single set of locks and conditional variables for each taskset, and they changed it to have a set of them for each thread of the taskset.

The big changes outside of taskset.c are just a simple reindentation caused by using the new taskset interface.