@jstebbins 2021-08-10
Make's sense.  Finer grained locking would improve utilization with many cores. I'll try making a local copy and using ```git diff -b``` to see if that makes the diffs more readable.