@jstebbins 2021-08-10
It looks like they also simplified the individual filters that use tasksets considerably.  Moved more of the management of tasks into the taskset code somehow.  That's a nice additional benefit.