@sr55 2021-08-10
Very limited initial testing on an AMD 5900x

**Source:** Big Buck Bunny 1080p
**Preset:** Very Fast 1080p (modified)
**Modifications:**
- Changed to x264, ultrafast preset.


### Test 1: Decomb/Interlace Detect Filters

**Before:**
```
[18:35:08] work: average encoding speed for job is 395.760223 fps
[18:41:30] work: average encoding speed for job is 397.273132 fps
[18:43:19] work: average encoding speed for job is 386.004303 fps
avg: 392
```

**After:**
```
[18:38:38] work: average encoding speed for job is 450.014832 fps
[18:45:00] work: average encoding speed for job is 431.441650 fps
[18:47:08] work: average encoding speed for job is 441.825378 fps
avg: 440
```

_~12% average increase._


### Test 2: No Filters

**Before:**
```
[18:55:43] work: average encoding speed for job is 637.191345 fps
[18:56:56] work: average encoding speed for job is 645.840454 fps
[18:58:05] work: average encoding speed for job is 645.515198 fps
avg: 642  (63% faster than without, pre-patch)
```

**After:**

```
[18:54:07] work: average encoding speed for job is 700.351074 fps
[18:59:10] work: average encoding speed for job is 687.457764 fps
[19:00:41] work: average encoding speed for job is 701.683289 fps
avg: 696  (58% faster than without, post-patch)

```
_~8% average increase._