@sr55 2021-08-10
Using more realistic settings seems to yield lesser results but an improvement all the same. 

PAL DVD with:
"Fast 1080p" preset yields around ~4%

Big Buck bunny with:
"H.265 MKV 1080p30" preset was  +/- 1% which I'd say is within margin of error so no improvement. 


I'll spend some time over the weekend testing a few different configs to see how this performs on lower end quad/hex core systems.