@jstebbins 2021-08-10
> Test 2: No Filters
~8% average increase.

Huh?  tasksets are only used with filters.  Should be no difference if no filters. Something smells like trout.