@sr55 2021-08-10
Yeh, I thought too.  I'll re-run the tests but that was across 3 runs it was fairly consistent.