@sr55 2021-08-10
Might be the FrameRate Shaper. BBB is 60fps and our presets are 30 PFR

Removing that is yielding identical results.