@jstebbins 2021-08-10
Framerate adjustments require literally no cpu and also don't use tasksets.  So something is stiff fishy.