@galad87 2021-09-01
@galinart I rebased this, but it would be better to wait until I (or someone else) implement #3810 so you don't have to fix the qsv part yet again.
I was thinking of adding a format filter at the end of the filters chain to convert to the encoder input format if the pipeline format is different (for example if the pipeline format is 10bit to get a better output from the colorspace filter and the encoder is 8bit).