@sr55 2021-11-06
I think it must have been PEBKAC on the FPS issue. Cannot reproduce that.

I've left it running for a few hours this morning testing out all the filters.  Results all came out as expected so I think the patch at least works as advertised.

On performance:

- Decomb/Comb Detect:  9~15% where the encoder isn't bottlenecked or CPU usage isn't 100% already
- NLMeans:  Within margin of error.
- Grayscale:  Within margin of error.