@sr55 2021-11-10
What do you think?  Commit and let it settle for a bit, or @jstebbins  do you want to have a closer look at the code first?