@jstebbins 2022-02-07
I've only looked at this close enough to understand *why* it's an improvement.  Haven't really looked hard for errors.  But if it seems to work, ship it!